+++ src/components/ModalTransactionEdit/ModalTransaction.module.scss
@@ -136,7 +136,10 @@
   font-size: 16px;
   line-height: 24px;
   color: #e0e0e0;
+  cursor: pointer;
 }
+
+
 
 .toggleText.activeExpense {
   color: #ff6596;
